requirements.txt
google-generativeai
sqlparse
python-dotenv
fastapi
uvicorn
